transport: create Client interface.
* a Client can instantiate any client transport service.
* InstallProtocol installs a Client for a given protocol,
  instead of just a UploadPackService.
* Renamed GitUploadPackService to UploadPackService.
* Adapt ssh and http clients to the new client interface.
