+++ plumbing/transport/common.go
@@ -1,7 +1,7 @@
 // Package clients includes the implementation for different transport protocols
 //
 // go-git needs the packfile and the refs of the repo. The
-// `NewGitUploadPackService` function returns an object that allows to
+// `NewUploadPackService` function returns an object that allows to
 // download them.
 //
 // go-git supports HTTP and SSH (see `Protocols`) for downloading the packfile
@@ -9,38 +9,79 @@
 // `InstallProtocol` below).
 //
 // Each protocol has its own implementation of
-// `NewGitUploadPackService`, but you should generally not use them
-// directly, use this package's `NewGitUploadPackService` instead.
+// `NewUploadPackService`, but you should generally not use them
+// directly, use this package's `NewUploadPackService` instead.
 package transport
 
 import (
+	"errors"
 	"fmt"
+	"net/url"
+	"regexp"
 
-	"gopkg.in/src-d/go-git.v4/plumbing/transport/common"
-	"gopkg.in/src-d/go-git.v4/plumbing/transport/http"
-	"gopkg.in/src-d/go-git.v4/plumbing/transport/ssh"
+	"gopkg.in/src-d/go-git.v4/plumbing"
 )
 
-// Protocols are the protocols supported by default.
-var Protocols = map[string]common.GitUploadPackServiceFactory{
-	"http":  http.NewGitUploadPackService,
-	"https": http.NewGitUploadPackService,
-	"ssh":   ssh.NewGitUploadPackService,
+var (
+	ErrRepositoryNotFound     = errors.New("repository not found")
+	ErrAuthorizationRequired  = errors.New("authorization required")
+	ErrEmptyUploadPackRequest = errors.New("empty git-upload-pack given")
+	ErrInvalidAuthMethod      = errors.New("invalid auth method")
+)
+
+const (
+	UploadPackServiceName = "git-upload-pack"
+)
+
+type Client interface {
+	Connect() error
+	Disconnect() error
+	SetAuth(AuthMethod) error
+	FetchPackService
 }
 
-// InstallProtocol adds or modifies an existing protocol.
-func InstallProtocol(scheme string, f common.GitUploadPackServiceFactory) {
-	Protocols[scheme] = f
+// ClientFactory instantiates a Client with given endpoint.
+type ClientFactory func(Endpoint) Client
+
+type AuthMethod interface {
+	fmt.Stringer
+	Name() string
 }
 
-// NewGitUploadPackService returns the appropriate upload pack service
-// among of the set of known protocols: HTTP, SSH. See `InstallProtocol`
-// to add or modify protocols.
-func NewGitUploadPackService(endpoint common.Endpoint) (common.GitUploadPackService, error) {
-	f, ok := Protocols[endpoint.Scheme]
-	if !ok {
-		return nil, fmt.Errorf("unsupported scheme %q", endpoint.Scheme)
+type Endpoint url.URL
+
+var (
+	isSchemeRegExp   = regexp.MustCompile("^[^:]+://")
+	scpLikeUrlRegExp = regexp.MustCompile("^(?P<user>[^@]+@)?(?P<host>[^:]+):/?(?P<path>.+)$")
+)
+
+func NewEndpoint(endpoint string) (Endpoint, error) {
+	endpoint = transformSCPLikeIfNeeded(endpoint)
+
+	u, err := url.Parse(endpoint)
+	if err != nil {
+		return Endpoint{}, plumbing.NewPermanentError(err)
 	}
 
-	return f(endpoint), nil
+	if !u.IsAbs() {
+		return Endpoint{}, plumbing.NewPermanentError(fmt.Errorf(
+			"invalid endpoint: %s", endpoint,
+		))
+	}
+
+	return Endpoint(*u), nil
+}
+
+func (e *Endpoint) String() string {
+	u := url.URL(*e)
+	return u.String()
+}
+
+func transformSCPLikeIfNeeded(endpoint string) string {
+	if !isSchemeRegExp.MatchString(endpoint) && scpLikeUrlRegExp.MatchString(endpoint) {
+		m := scpLikeUrlRegExp.FindStringSubmatch(endpoint)
+		return fmt.Sprintf("ssh://%s%s/%s", m[1], m[2], m[3])
+	}
+
+	return endpoint
 }
