--- plumbing/transport/http/common_test.go
@@ -1,51 +1,0 @@
-package http
-
-import (
-	"net/http"
-	"testing"
-
-	. "gopkg.in/check.v1"
-)
-
-func Test(t *testing.T) { TestingT(t) }
-
-type SuiteCommon struct{}
-
-var _ = Suite(&SuiteCommon{})
-
-func (s *SuiteCommon) TestNewBasicAuth(c *C) {
-	a := NewBasicAuth("foo", "qux")
-
-	c.Assert(a.Name(), Equals, "http-basic-auth")
-	c.Assert(a.String(), Equals, "http-basic-auth - foo:*******")
-}
-
-func (s *SuiteCommon) TestNewErrOK(c *C) {
-	res := &http.Response{StatusCode: http.StatusOK}
-	err := NewErr(res)
-	c.Assert(err, IsNil)
-}
-
-func (s *SuiteCommon) TestNewErrUnauthorized(c *C) {
-	s.testNewHTTPError(c, http.StatusUnauthorized, "authorization required")
-}
-
-func (s *SuiteCommon) TestNewErrNotFound(c *C) {
-	s.testNewHTTPError(c, http.StatusNotFound, "repository not found")
-}
-
-func (s *SuiteCommon) TestNewHTTPError40x(c *C) {
-	s.testNewHTTPError(c, http.StatusPaymentRequired, "unexpected client error.*")
-}
-
-func (s *SuiteCommon) testNewHTTPError(c *C, code int, msg string) {
-	req, _ := http.NewRequest("GET", "foo", nil)
-	res := &http.Response{
-		StatusCode: code,
-		Request:    req,
-	}
-
-	err := NewErr(res)
-	c.Assert(err, NotNil)
-	c.Assert(err, ErrorMatches, msg)
-}
