+++ plumbing/transport/ssh/auth_method.go
@@ -7,14 +7,15 @@ import (
 
 	"golang.org/x/crypto/ssh"
 	"golang.org/x/crypto/ssh/agent"
-	"gopkg.in/src-d/go-git.v4/plumbing/transport/common"
+
+	"gopkg.in/src-d/go-git.v4/plumbing/transport"
 )
 
 // AuthMethod is the interface all auth methods for the ssh client
 // must implement. The clientConfig method returns the ssh client
 // configuration needed to establish an ssh connection.
 type AuthMethod interface {
-	common.AuthMethod
+	transport.AuthMethod
 	clientConfig() *ssh.ClientConfig
 }
