+++ plumbing/transport/ssh/auth_method_test.go
@@ -2,12 +2,9 @@ package ssh
 
 import (
 	"fmt"
-	"testing"
 
 	. "gopkg.in/check.v1"
 )
-
-func Test(t *testing.T) { TestingT(t) }
 
 type SuiteCommon struct{}
